+++ admin/supprimer.php
@@ -4,7 +4,7 @@ include '../config.php';
 
 $id = $_GET['id'];
 
-$sql = 'DELETE FROM portfolio WHERE port_id=:id';
+$sql = 'DELETE FROM portfolio WHERE port_id=:id AND admin_id=1';
 
 $stmt = $db->prepare($sql);
